ui/requirements.txt
asgiref==3.2.10
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
Django==3.0.6
django-extensions==2.2.9
django-filter==2.2.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
